+++ snowddl/app/base.py
@@ -722,7 +722,7 @@ class BaseApp:
         self.logger.info(f"Executed {len(engine.executed_ddl)} DDL queries, Suggested {len(engine.suggested_ddl)} DDL queries")
 
     def output_engine_warnings(self, engine: SnowDDLEngine):
-        for object_type, object_names in engine.intention_cache.invalid_name_warning.items():
+        for object_type, object_names in engine.intention_cache.object_name_warning.items():
             for name in object_names:
                 self.logger.warning(
                     f"Detected {object_type.name} with name [{name}] "