+++ snowddl/resolver/abc_resolver.py
@@ -95,8 +95,8 @@ class AbstractResolver(ABC):
             if object_full_name in self.blueprints:
                 continue
 
-            # Parent object is going to be dropped
-            if self.engine.intention_cache.check_parent_drop_intention(self.object_type, object_full_name):
+            # Another object is going to be dropped, which implicitly drops this object
+            if self._check_implicit_drop_intention(object_full_name):
                 continue
 
             tasks[object_full_name] = (self.drop_object, self.existing_objects[object_full_name])
@@ -107,14 +107,21 @@ class AbstractResolver(ABC):
         tasks = {}
 
         # Drop all existing objects
-        for full_name in sorted(self.existing_objects):
-            # Parent object is going to be dropped
-            if self.engine.intention_cache.check_parent_drop_intention(self.object_type, full_name):
+        for object_full_name in sorted(self.existing_objects):
+            # Another object is going to be dropped, which implicitly drops this object
+            if self._check_implicit_drop_intention(object_full_name):
                 continue
 
-            tasks[full_name] = (self.drop_object, self.existing_objects[full_name])
+            tasks[object_full_name] = (self.drop_object, self.existing_objects[object_full_name])
 
         self._process_tasks(tasks)
+
+    def _check_implicit_drop_intention(self, object_full_name: str) -> bool:
+        # Dropping or replacing any parent object implicitly drops this object
+        if self.engine.intention_cache.check_parent_object_drop_intention(self.object_type, object_full_name):
+            return True
+
+        return False
 
     def _process_tasks(self, tasks):
         futures = {}
@@ -128,11 +135,8 @@ class AbstractResolver(ABC):
             try:
                 result = f.result()
 
-                if result == ResolveResult.REPLACE:
-                    self.engine.intention_cache.add_replace_intention(self.object_type, full_name)
-
-                if result == ResolveResult.DROP:
-                    self.engine.intention_cache.add_drop_intention(self.object_type, full_name)
+                if result in (ResolveResult.REPLACE, ResolveResult.DROP):
+                    self.engine.intention_cache.add_object_drop_intention(self.object_type, full_name)
 
                 if result == ResolveResult.NOCHANGE:
                     self.engine.logger.debug(f"Resolved {self.object_type.name} [{full_name}]: {result.value}")