+++ snowddl/resolver/abc_schema_object_resolver.py
@@ -28,8 +28,8 @@ class AbstractSchemaObjectResolver(AbstractResolver):
             if object_full_name in self.blueprints:
                 continue
 
-            # Parent object is going to be dropped
-            if self.engine.intention_cache.check_parent_drop_intention(self.object_type, object_full_name):
+            # Another object is going to be dropped, which implicitly drops this object
+            if self._check_implicit_drop_intention(object_full_name):
                 continue
 
             schema_full_name = ".".join(object_full_name.split(".")[:2])