+++ snowddl/resolver/schema.py
@@ -108,7 +108,7 @@ class SchemaResolver(AbstractResolver):
                 continue
 
             # Parent object is going to be dropped
-            if self.engine.intention_cache.check_parent_drop_intention(self.object_type, schema_full_name):
+            if self.engine.intention_cache.check_parent_object_drop_intention(self.object_type, schema_full_name):
                 continue
 
             database_full_name = ".".join(schema_full_name.split(".")[:1])