+++ snowddl/resolver/user_role.py
@@ -28,7 +28,7 @@ class UserRoleResolver(AbstractRoleResolver):
             try:
                 grant_name = build_grant_name_ident(self.config.env_prefix, r["name"], object_type)
             except (KeyError, ValueError):
-                self.engine.intention_cache.add_invalid_name_warning(object_type, r["name"])
+                self.engine.intention_cache.add_object_name_warning(object_type, r["name"])
                 continue
 
             grants.append(